Add spinner while chime is loading on folder page
If a user (or cypress) tries to create a folder before the chime is
loaded, the creation will fail silently.

This adds a simple loader while witing for the chime information.

When `reloadChime` is called, `isReady` will be false. When the
promise settled (either fulfilled or rejected), `isReady` will be true.