get draggable component working
the updated `<Draggable>` component has a different API.
It doesn't want to rerender after an update (maybe because the question_responses don't have a real id for an itemKey, so I'm forcing a rerendering by settings a :key on the whole component.

There's gotta be a better way to do this.