migrate tests and helpers to ts
These tests were breaking with the move to `vite` because as part of the move, we needed to remove `type: module` from package.json (at least until they fix the issue).

This means any `.js` file will be interpretted as a `common js` file, so it expects `require` and not `import`.

Instead of just renaming all of these to `.mjs` files, we change them to `.ts` and fix any type issues.