handle case where topic1 is substring of topic2
We need to sort the topics and verbs from largest to smallest so that shorter substrings aren't removed first.

For example, if we have text `James is great. So is James Bond.` This would generate topics: `['James', 'James Bond']`. If we don't sort first, when we go to remove topics, we'll:

1)  remove `James` from text => `is great. So is Bond` // leaves `Bond`
2) remove `James Bond` from text =>  same as above, since `James Bond` doesn't match.

Then when generating terms from the scrubbed text, it'll include `Bond` giving an extra `Bond` in our final list: `['James', 'James Bond', ..., 'Bond', ...]

Sorting the largest words first will fix this issue. This also adds a test to avoid regressions.