change from classic `yarn` to `npm`
classic `yarn` (v1) is deprecated. When attempting to update some packages, `yarn install` fails with mysterious errors. The options are to workaround by resolving to some outdated depdencies, update to a newer version of yarn (v4), or change to a different  package manager (npm or pnpm).

updating to yarn or changing to npm both involve enabling corepack. This may make sense in the future.

But, for now, changing to use the default npm package manager seems fine.

This adds the `package-lock.json` for `npm` and updates `deploy.php` to use `npm`.

The contents of `deploy.php` is shuffled to align with the order of the Laravel recipe. Commented out tasks are removed.