fix: question move to end after editing
This regression was introduced in #945, where we moved a question to the end of a folder when current folder id isn't the destination folder id.

To determine if we need to move the question, we compare the current folder id and the destination folder id.

However, we have a type mismatch because we're getting the current folder id from the route.
```
// $currentFolderId is a string
$currentFolderId = $req->route('folder_id');

// $destFolderId is an int
$destFolderId = $req->get('folder_id');
```

Thus, currentFolderId will always be different from destFolderId.

The solution is to cast the currentFolderId as an int so that they can be properly compared.